+++ libsolidity/interface/SMTSolverCommand.cpp
@@ -30,10 +30,6 @@
 #include <boost/filesystem/fstream.hpp>
 #include <boost/process.hpp>
 
-using solidity::langutil::InternalCompilerError;
-using solidity::util::errinfo_comment;
-
-
 namespace solidity::frontend
 {
 
@@ -51,7 +47,23 @@ void SMTSolverCommand::setEldarica(std::optional<unsigned int> timeoutInMillisec
 		m_arguments.emplace_back("-ssol");
 }
 
-ReadCallback::Result SMTSolverCommand::solve(std::string const& _kind, std::string const& _query)
+void SMTSolverCommand::setCvc5(std::optional<unsigned int> timeoutInMilliseconds)
+{
+	m_arguments.clear();
+	m_solverCmd = "cvc5";
+	if (timeoutInMilliseconds)
+	{
+		m_arguments.push_back("--tlimit-per");
+		m_arguments.push_back(std::to_string(timeoutInMilliseconds.value()));
+	}
+	else
+	{
+		m_arguments.push_back("--rlimit");
+		m_arguments.push_back(std::to_string(12000));
+	}
+}
+
+ReadCallback::Result SMTSolverCommand::solve(std::string const& _kind, std::string const& _query) const
 {
 	try
 	{
@@ -68,35 +80,35 @@ ReadCallback::Result SMTSolverCommand::solve(std::string const& _kind, std::stri
 		auto queryFile = boost::filesystem::ofstream(queryFileName);
 		queryFile << _query << std::flush;
 
-		auto eldBin = boost::process::search_path(m_solverCmd);
+		auto solverBin = boost::process::search_path(m_solverCmd);
 
-		if (eldBin.empty())
+		if (solverBin.empty())
 			return ReadCallback::Result{false, m_solverCmd + " binary not found."};
 
 		auto args = m_arguments;
 		args.push_back(queryFileName.string());
 
 		boost::process::ipstream pipe;
-		boost::process::child eld(
-			eldBin,
+		boost::process::child solverProcess(
+			solverBin,
 			args,
 			boost::process::std_out > pipe,
-			boost::process::std_err >boost::process::null
+			boost::process::std_err > boost::process::null
 		);
 
 		std::vector<std::string> data;
 		std::string line;
-		while (eld.running() && std::getline(pipe, line))
+		while (solverProcess.running() && std::getline(pipe, line))
 			if (!line.empty())
 				data.push_back(line);
 
-		eld.wait();
+		solverProcess.wait();
 
 		return ReadCallback::Result{true, boost::join(data, "\n")};
 	}
 	catch (...)
 	{
-		return ReadCallback::Result{false, "Unknown exception in SMTQuery callback: " + boost::current_exception_diagnostic_information()};
+		return ReadCallback::Result{false, "Exception in SMTQuery callback: " + boost::current_exception_diagnostic_information()};
 	}
 }