+++ test/libsolidity/MemoryGuardTest.cpp
@@ -73,14 +73,14 @@ TestCase::TestResult MemoryGuardTest::run(std::ostream& _stream, std::string con
 		}
 
 		auto handleObject = [&](std::string const& _kind, Object const& _object) {
-			m_obtainedResult += contractName + "(" + _kind + ") " + (FunctionCallFinder::run(
+			m_obtainedResult += contractName + "(" + _kind + ") " + (findFunctionCalls(
 				*_object.code,
-				"memoryguard"_yulstring
+				"memoryguard"_yulname
 			).empty() ? "false" : "true") + "\n";
 		};
 		handleObject("creation", *object);
 		size_t deployedIndex = object->subIndexByName.at(
-			YulString(IRNames::deployedObject(compiler().contractDefinition(contractName)))
+			IRNames::deployedObject(compiler().contractDefinition(contractName))
 		);
 		handleObject("runtime", dynamic_cast<Object const&>(*object->subObjects[deployedIndex]));
 	}