+++ test/tools/fuzzer_common.cpp
@@ -49,7 +49,8 @@ static std::vector<EVMVersion> s_evmVersions = {
 	EVMVersion::istanbul(),
 	EVMVersion::berlin(),
 	EVMVersion::london(),
-	EVMVersion::paris()
+	EVMVersion::paris(),
+	EVMVersion::prague()
 };
 
 void FuzzerUtil::testCompilerJsonInterface(std::string const& _input, bool _optimize, bool _quiet)
@@ -126,12 +127,6 @@ void FuzzerUtil::testCompiler(
 	try
 	{
 		compiler.compile();
-	}
-	catch (Error const&)
-	{
-	}
-	catch (FatalError const&)
-	{
 	}
 	catch (UnimplementedFeatureError const&)
 	{